backend/pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fiona>=1.10.1",
    "flask-cors>=5.0.0",
    "flask-sqlalchemy>=3.1.1",
    "flask>=3.0.3",
    "geoalchemy2>=0.15.2",
    "geojson>=3.1.0",
    "geopandas>=1.0.1",
    "gunicorn[gevent]>=23.0.0",
    "openai>=1.51.0",
    "psycopg2-binary>=2.9.9",
    "requests>=2.32.3",
    "ruff>=0.6.9",
]
